- clean up xml-related memory handling - std::min is pickier about types...

+++ apt-pkg/contrib/hashes.cc
@@ -30,8 +30,8 @@ bool Hashes::AddFD(int Fd,unsigned long Size)
    int Res = 0;
    while (Size != 0)
    {
-      Res = read(Fd,Buf,std::min(Size,sizeof(Buf)));
-      if (Res < 0 || (unsigned)Res != std::min(Size,sizeof(Buf)))
+      Res = read(Fd,Buf,std::min(Size,(unsigned long)sizeof(Buf)));
+      if (Res < 0 || (unsigned)Res != std::min(Size,(unsigned long)sizeof(Buf)))
 	 return false;
       Size -= Res;
       MD5.Add(Buf,Res);
